porsline_config/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
defusedxml==0.7.1
Django==4.1.7
django-cors-headers==4.0.0
django-filter==23.2
django-templated-mail==1.1.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-nested-routers==0.93.4
drf-yasg==1.21.5
exceptiongroup==1.1.1
frozenlist==1.3.3
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
isodate==0.6.1
itypes==1.2.0
jdatetime==4.1.1
Jinja2==3.1.2
kombu==5.2.4
lxml==4.9.2
MarkupSafe==2.1.2
model-bakery==1.11.0
multidict==6.0.4
oauthlib==3.2.2
packaging==23.0
Pillow==9.4.0
platformdirs==3.5.1
pluggy==1.0.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.6.0
pytest==7.3.1
pytest-django==4.5.2
pytest-dotenv==0.5.2
python-decouple==3.8
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2022.7.1
redis==4.5.5
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
zeep==4.2.1
